+++ examples/vt_version_3_object_pool/main.cpp
@@ -1,4 +1,5 @@
 #include "can_general_parameter_group_numbers.hpp"
+#include "can_hardware_interface.hpp"
 #include "can_network_manager.hpp"
 #include "can_partnered_control_function.hpp"
 #include "iop_file_interface.hpp"
@@ -16,6 +17,7 @@ static std::shared_ptr<isobus::VirtualTerminalClient> TestVirtualTerminalClient 
 std::vector<isobus::NAMEFilter> vtNameFilters;
 const isobus::NAMEFilter testFilter(isobus::NAME::NAMEParameters::FunctionCode, static_cast<std::uint8_t>(isobus::NAME::Function::VirtualTerminal));
 static std::vector<std::uint8_t> testPool;
+static SocketCANInterface canDriver("can0");
 
 using namespace std;
 
@@ -42,7 +44,7 @@ void raw_can_glue(isobus::HardwareInterfaceCANFrame &rawFrame, void *parentPoint
 void setup()
 {
 	CANHardwareInterface::set_number_of_can_channels(1);
-	CANHardwareInterface::assign_can_channel_frame_handler(0, "can0");
+	CANHardwareInterface::assign_can_channel_frame_handler(0, &canDriver);
 	CANHardwareInterface::start();
 
 	CANHardwareInterface::add_can_lib_update_callback(update_CAN_network, nullptr);
