+++ hardware_integration/src/can_hardware_interface.cpp
@@ -7,7 +7,7 @@
 ///
 /// @copyright 2022 Adrian Del Grosso
 //================================================================================================
-#include "socket_can_interface.hpp"
+#include "can_hardware_interface.hpp"
 #include "system_timing.hpp"
 
 #include <linux/can.h>
@@ -42,191 +42,6 @@ bool isobus::send_can_message_to_hardware(HardwareInterfaceCANFrame frame)
 	return CANHardwareInterface::transmit_can_message(frame);
 }
 
-CANHardwareInterface::SocketCANFrameHandler::SocketCANFrameHandler(const std::string deviceName) :
-  pCANDevice(new sockaddr_can),
-  name(deviceName),
-  fileDescriptor(-1)
-{
-	if (nullptr != pCANDevice)
-	{
-		memset(pCANDevice, 0, sizeof(struct sockaddr_can));
-	}
-}
-
-CANHardwareInterface::SocketCANFrameHandler::~SocketCANFrameHandler()
-{
-	close();
-
-	if (nullptr != pCANDevice)
-	{
-		delete pCANDevice;
-		pCANDevice = nullptr;
-	}
-}
-
-bool CANHardwareInterface::SocketCANFrameHandler::get_is_valid() const
-{
-	return (-1 != fileDescriptor);
-}
-
-std::string CANHardwareInterface::SocketCANFrameHandler::get_device_name() const
-{
-	return name;
-}
-
-void CANHardwareInterface::SocketCANFrameHandler::close()
-{
-	::close(fileDescriptor);
-	fileDescriptor = -1;
-}
-
-void CANHardwareInterface::SocketCANFrameHandler::open()
-{
-	fileDescriptor = socket(PF_CAN, SOCK_RAW, CAN_RAW);
-
-	if (fileDescriptor >= 0)
-	{
-		struct ifreq interfaceRequestStructure;
-		const int RECEIVE_OWN_MESSAGES = 0;
-		const int DROP_MONITOR = 1;
-		const int TIMESTAMPING = 0x58;
-		const int TIMESTAMP = 1;
-		memset(&interfaceRequestStructure, 0, sizeof(interfaceRequestStructure));
-		strncpy(interfaceRequestStructure.ifr_name, name.c_str(), sizeof(interfaceRequestStructure.ifr_name));
-		setsockopt(fileDescriptor, SOL_CAN_RAW, CAN_RAW_RECV_OWN_MSGS, &RECEIVE_OWN_MESSAGES, sizeof(RECEIVE_OWN_MESSAGES));
-		setsockopt(fileDescriptor, SOL_SOCKET, SO_RXQ_OVFL, &DROP_MONITOR, sizeof(DROP_MONITOR));
-
-		if (setsockopt(fileDescriptor, SOL_SOCKET, SO_TIMESTAMPING, &TIMESTAMPING, sizeof(TIMESTAMPING)) < 0)
-		{
-			setsockopt(fileDescriptor, SOL_SOCKET, SO_TIMESTAMP, &TIMESTAMP, sizeof(TIMESTAMP));
-		}
-
-		if (ioctl(fileDescriptor, SIOCGIFINDEX, &interfaceRequestStructure) >= 0)
-		{
-			memset(pCANDevice, 0, sizeof(sockaddr_can));
-			pCANDevice->can_family = AF_CAN;
-			pCANDevice->can_ifindex = interfaceRequestStructure.ifr_ifindex;
-
-			if (bind(fileDescriptor, (struct sockaddr *)pCANDevice, sizeof(struct sockaddr)) < 0)
-			{
-				::close(fileDescriptor);
-				fileDescriptor = -1;
-			}
-		}
-		else
-		{
-			::close(fileDescriptor);
-			fileDescriptor = -1;
-		}
-	}
-	else
-	{
-		::close(fileDescriptor);
-		fileDescriptor = -1;
-	}
-}
-
-bool CANHardwareInterface::SocketCANFrameHandler::read_frame(isobus::HardwareInterfaceCANFrame &canFrame)
-{
-	struct pollfd pollingFileDescriptor;
-	bool retVal = false;
-
-	pollingFileDescriptor.fd = fileDescriptor;
-	pollingFileDescriptor.events = POLLIN;
-	pollingFileDescriptor.revents = 0;
-
-	if (1 == poll(&pollingFileDescriptor, 1, 100))
-	{
-		canFrame.timestamp_us = std::numeric_limits<std::uint64_t>::max();
-		struct can_frame txFrame;
-		struct msghdr message;
-		struct iovec segment;
-
-		char lControlMessage[CMSG_SPACE(sizeof(struct timeval) + (3 * sizeof(struct timespec)) + sizeof(std::uint32_t))];
-
-		segment.iov_base = &txFrame;
-		segment.iov_len = sizeof(struct can_frame);
-		message.msg_iov = &segment;
-		message.msg_iovlen = 1;
-		message.msg_control = &lControlMessage;
-		message.msg_controllen = sizeof(lControlMessage);
-		message.msg_name = pCANDevice;
-		message.msg_namelen = sizeof(struct sockaddr_can);
-		message.msg_flags = 0;
-
-		if (recvmsg(fileDescriptor, &message, 0) > 0)
-		{
-			if (0 == (txFrame.can_id & CAN_ERR_FLAG))
-			{
-				if (0 != (txFrame.can_id & CAN_EFF_FLAG))
-				{
-					canFrame.identifier = (txFrame.can_id & CAN_EFF_MASK);
-					canFrame.isExtendedFrame = true;
-				}
-				else
-				{
-					canFrame.identifier = (txFrame.can_id & CAN_SFF_MASK);
-					canFrame.isExtendedFrame = false;
-				}
-				canFrame.dataLength = txFrame.can_dlc;
-				memset(canFrame.data, 0, sizeof(canFrame.data));
-				memcpy(canFrame.data, txFrame.data, canFrame.dataLength);
-
-				for (struct cmsghdr *pControlMessage = CMSG_FIRSTHDR(&message); (nullptr != pControlMessage) && (SOL_SOCKET == pControlMessage->cmsg_level); pControlMessage = CMSG_NXTHDR(&message, pControlMessage))
-				{
-					switch (pControlMessage->cmsg_type)
-					{
-						case SO_TIMESTAMP:
-						{
-							struct timeval *time = (struct timeval *)CMSG_DATA(pControlMessage);
-
-							if (std::numeric_limits<std::uint64_t>::max() == canFrame.timestamp_us)
-							{
-								canFrame.timestamp_us = static_cast<std::uint64_t>(time->tv_usec) + (static_cast<std::uint64_t>(time->tv_sec) * 1000000);
-							}
-						}
-						break;
-
-						case SO_TIMESTAMPING:
-						{
-							struct timespec *time = (struct timespec *)(CMSG_DATA(pControlMessage));
-							canFrame.timestamp_us = (static_cast<std::uint64_t>(time[2].tv_nsec) / 1000) + (static_cast<std::uint64_t>(time[2].tv_sec) * 1000000);
-						}
-						break;
-					}
-				}
-				retVal = true;
-			}
-		}
-	}
-	else if (pollingFileDescriptor.revents & (POLLERR | POLLHUP))
-	{
-		close();
-	}
-	return retVal;
-}
-
-bool CANHardwareInterface::SocketCANFrameHandler::write_frame(const isobus::HardwareInterfaceCANFrame &canFrame)
-{
-	struct can_frame txFrame;
-	bool retVal = false;
-
-	txFrame.can_id = canFrame.identifier;
-	txFrame.can_dlc = canFrame.dataLength;
-	memcpy(txFrame.data, canFrame.data, canFrame.dataLength);
-
-	if (canFrame.isExtendedFrame)
-	{
-		txFrame.can_id |= CAN_EFF_FLAG;
-	}
-
-	if (write(fileDescriptor, &txFrame, sizeof(struct can_frame)) > 0)
-	{
-		retVal = true;
-	}
-	return retVal;
-}
-
 CANHardwareInterface::RawCanMessageCallbackInfo::RawCanMessageCallbackInfo() :
   callback(nullptr),
   parent(nullptr)
@@ -258,7 +73,7 @@ CANHardwareInterface::~CANHardwareInterface()
 	set_number_of_can_channels(0);
 }
 
-bool CANHardwareInterface::assign_can_channel_frame_handler(std::uint8_t aCANChannel, std::string deviceName)
+bool CANHardwareInterface::assign_can_channel_frame_handler(std::uint8_t aCANChannel, CANHardwarePlugin *driver)
 {
 	bool retVal = false;
 
@@ -267,21 +82,11 @@ bool CANHardwareInterface::assign_can_channel_frame_handler(std::uint8_t aCANCha
 		if ((!threadsStarted) &&
 		    (aCANChannel < hardwareChannels.size()))
 		{
-			if (nullptr == hardwareChannels[aCANChannel]->frameHandler)
+			if ((nullptr == hardwareChannels[aCANChannel]->frameHandler) ||
+			    (driver == hardwareChannels[aCANChannel]->frameHandler))
 			{
 				retVal = true;
-				hardwareChannels[aCANChannel]->frameHandler = new SocketCANFrameHandler(deviceName);
-			}
-			else
-			{
-				retVal = true;
-
-				if (hardwareChannels[aCANChannel]->frameHandler->get_device_name() != deviceName)
-				{
-					hardwareChannels[aCANChannel]->frameHandler->close();
-					delete hardwareChannels[aCANChannel]->frameHandler;
-					hardwareChannels[aCANChannel]->frameHandler = new SocketCANFrameHandler(deviceName);
-				}
+				hardwareChannels[aCANChannel]->frameHandler = driver;
 			}
 		}
 		hardwareChannelsMutex.unlock();
@@ -323,7 +128,6 @@ bool CANHardwareInterface::set_number_of_can_channels(uint8_t aValue)
 					{
 						pCANHardware->frameHandler->close();
 					}
-					delete pCANHardware->frameHandler;
 				}
 
 				delete pCANHardware;
