+++ test/address_claim_test.cpp
@@ -1,6 +1,7 @@
 #include <gtest/gtest.h>
 
 #include "can_constants.hpp"
+#include "can_hardware_interface.hpp"
 #include "can_internal_control_function.hpp"
 #include "can_network_manager.hpp"
 #include "socket_can_interface.hpp"
@@ -35,8 +36,9 @@ TEST(ADDRESS_CLAIM_TESTS, NAMETests)
 
 TEST(ADDRESS_CLAIM_TESTS, AddressClaiming)
 {
+	SocketCANInterface canDriver("can0");
 	CANHardwareInterface::set_number_of_can_channels(1);
-	CANHardwareInterface::assign_can_channel_frame_handler(0, "vcan0");
+	CANHardwareInterface::assign_can_channel_frame_handler(0, &canDriver);
 	CANHardwareInterface::start();
 
 	CANHardwareInterface::add_can_lib_update_callback(update_CAN_network, nullptr);
